benchmarks/requirements.txt
# Diffusion benchmarking only. Isolated from svg-engine/StarVector.
diffusers
transformers
accelerate
safetensors
torch
torchvision
pillow
